src/kernns/requirements.txt
Cython
flask
annoy
msgpack
python-rocksdb
scipy
